scanner (tokenizer) is working, and basic parser implemented
